+++ examples/clap.rs
@@ -5,6 +5,7 @@ extern crate log;
 extern crate stderrlog;
 
 use clap::{Arg, App};
+use std::str::FromStr;
 
 fn main() {
     let m = App::new("stderrlog example")
@@ -16,15 +17,30 @@ fn main() {
         .arg(Arg::with_name("quiet")
              .short("q")
              .help("Silence all output"))
+        .arg(Arg::with_name("timestamp")
+             .short("t")
+             .help("prepend log lines with a timestamp")
+             .takes_value(true)
+             .possible_values(&["none", "sec", "ms", "ns"]))
         .get_matches();
 
     let verbose = m.occurrences_of("verbosity") as usize;
     let quiet = m.is_present("quiet");
+    let ts = m.value_of("timestamp").map(|v| {
+        stderrlog::Timestamp::from_str(v).unwrap_or_else(|_| {
+            clap::Error {
+                message: "invalid value for 'timestamp'".into(),
+                kind: clap::ErrorKind::InvalidValue,
+                info: None,
+            }.exit()
+        })
+    }).unwrap_or(stderrlog::Timestamp::Off);
 
     stderrlog::new()
         .module(module_path!())
         .quiet(quiet)
         .verbosity(verbose)
+        .timestamp(ts)
         .init()
         .unwrap();
     trace!("trace message");